week_2/day_5/daily_challenge/Restaurant_Menu_Manager_api/pyproject.toml
[project]
name = "restaurant-menu-manager-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
